refactor(src): rename `useTableDragSelect`
